+++ assets/js/modules/pagespeed-insights/components/common/LabReportMetrics.js
@@ -20,49 +20,21 @@
  * External dependencies
  */
 import PropTypes from 'prop-types';
-import classnames from 'classnames';
 
 /**
  * WordPress dependencies
  */
-import { useCallback } from '@wordpress/element';
 import { __, _x } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
-import API from 'googlesitekit-api';
-import Data from 'googlesitekit-data';
-import { STORE_NAME as CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
 import ReportMetric from './ReportMetric';
-import ReportDetailsLink from './ReportDetailsLink';
 import MetricsLearnMoreLink from './MetricsLearnMoreLink';
 import { getScoreCategory } from '../../util';
-import Link from '../../../../components/Link';
 import ErrorText from '../../../../components/ErrorText';
-import {
-	STORE_NAME,
-	STRATEGY_MOBILE,
-	STRATEGY_DESKTOP,
-} from '../../datastore/constants';
-
-const { useSelect, useDispatch } = Data;
 
 export default function LabReportMetrics( { data, error } ) {
-	const referenceURL = useSelect( ( select ) => select( CORE_SITE ).getCurrentReferenceURL() );
-
-	const { invalidateResolution } = useDispatch( STORE_NAME );
-	const updateReport = useCallback( async ( event ) => {
-		event.preventDefault();
-
-		// Invalidate the PageSpeed API request caches.
-		await API.invalidateCache( 'modules', 'pagespeed-insights', 'pagespeed' );
-
-		// Invalidate the cached resolver.
-		invalidateResolution( 'getReport', [ referenceURL, STRATEGY_DESKTOP ] );
-		invalidateResolution( 'getReport', [ referenceURL, STRATEGY_MOBILE ] );
-	}, [ invalidateResolution, referenceURL ] );
-
 	const totalBlockingTime = data?.lighthouseResult?.audits?.[ 'total-blocking-time' ];
 	const largestContentfulPaint = data?.lighthouseResult?.audits?.[ 'largest-contentful-paint' ];
 	const cumulativeLayoutShift = data?.lighthouseResult?.audits?.[ 'cumulative-layout-shift' ];
@@ -90,12 +62,7 @@ export default function LabReportMetrics( { data, error } ) {
 					<MetricsLearnMoreLink />
 				</p>
 			</div>
-			<table
-				className={ classnames(
-					'googlesitekit-table',
-					'googlesitekit-table--with-list'
-				) }
-			>
+			<table className="googlesitekit-table googlesitekit-table--with-list">
 				<thead>
 					<tr>
 						<th>
@@ -124,15 +91,10 @@ export default function LabReportMetrics( { data, error } ) {
 						description={ __( 'How stable the elements on the page are', 'google-site-kit' ) }
 						displayValue={ cumulativeLayoutShift.displayValue }
 						category={ getScoreCategory( cumulativeLayoutShift.score ) }
+						isLast
 					/>
 				</tbody>
 			</table>
-			<div className="googlesitekit-pagespeed-report__row googlesitekit-pagespeed-report__row--last">
-				<Link onClick={ updateReport }>
-					{ __( 'Run test again', 'google-site-kit' ) }
-				</Link>
-				<ReportDetailsLink />
-			</div>
 		</div>
 	);
 }
