+++ assets/js/modules/pagespeed-insights/components/common/ReportMetric.js
@@ -37,6 +37,7 @@ export default function ReportMetric( {
 	description,
 	displayValue,
 	category,
+	isLast,
 } ) {
 	// Normalize the category case.
 	category = category.toLowerCase();
@@ -45,7 +46,10 @@ export default function ReportMetric( {
 		<tr
 			className={ classnames(
 				'googlesitekit-pagespeed-report__row',
-				'googlesitekit-pagespeed-report-metric'
+				'googlesitekit-pagespeed-report-metric',
+				{
+					'googlesitekit-pagespeed-report__row--last': isLast,
+				}
 			) }
 		>
 			<td>
@@ -57,12 +61,7 @@ export default function ReportMetric( {
 				</div>
 			</td>
 
-			<td
-				className={ classnames(
-					'googlesitekit-pagespeed-report-metric-value',
-					`googlesitekit-pagespeed-report-metric--${ category.toLowerCase() }`
-				) }
-			>
+			<td className={ `googlesitekit-pagespeed-report-metric-value googlesitekit-pagespeed-report-metric--${ category }` }>
 				<div className="googlesitekit-pagespeed-report-metric-value-container">
 					<div className="googlesitekit-pagespeed-report-metric-value__display-value">
 						{ displayValue }
@@ -83,4 +82,5 @@ ReportMetric.propTypes = {
 	description: PropTypes.string.isRequired,
 	displayValue: PropTypes.string.isRequired,
 	category: PropTypes.string.isRequired,
+	isLast: PropTypes.bool,
 };
